Day 7 | Updated libs to 2.5.0 and cleared TODOs

+++ implementation/src/main/kotlin/io/github/tomplum/aoc/computer/MultiplyInstruction.kt
@@ -8,5 +8,5 @@ class MultiplyInstruction(private val instruction: String) {
         .removeSuffix(")")
         .split(",")
         .map { value -> value.toInt() }
-        .product()
+        .product() as Int
 }
